frontend_streamlit/pyproject.toml
[tool.poetry]
name = "frontend-streamlit"
version = "0.1.0"
description = ""
authors = ["Adam Górski <adamsebastiangorski@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">3.9.7,<4.0"
streamlit = "^1.22.0"
folium = "^0.14.0"
streamlit-folium = "^0.11.1"
requests = "^2.29.0"
pandas = "^2.0.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
